Adding function for admin to view employee attendance
