+++ include/attendance_display.h
@@ -1,0 +1,1 @@
+void view_attendance(int);
